requirements.txt
acme==4.0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anticaptchaofficial==1.0.64
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==4.2.1
blinker==1.4
boto3==1.35.99
botocore==1.35.99
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
cattrs==24.1.3
celery==5.5.2
certbot==4.0.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
colourmap==1.1.21
ConfigArgParse==1.7
configobj==5.0.9
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.20
django-allauth==65.7.0
django-appconf==1.1.0
django-compressor==4.5.1
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-htmlmin==0.11.0
django-rest-framework==0.1.0
django-storages==1.14.4
django-user-agents==0.4.0
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fake-headers==1.0.2
fake-useragent==2.1.0
ffmpeg-python==0.2.0
fonttools==4.57.0
frozenlist==1.5.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
googletrends==0.1.2
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpie==3.2.4
httplib2==0.22.0
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
joblib==1.5.0
josepy==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaitaistruct==0.10
kiwisolver==1.4.8
kombu==5.5.3
lxml==4.9.3
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
multidict==6.1.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
numpy==2.2.5
olefile==0.47
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
pillow==11.1.0
platformdirs==4.3.7
playwright==1.52.0
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==5.9.5
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydivert==2.1.0
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyRFC3339==2.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.0
pytrends==4.9.2
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.13.0
ratelimit==2.2.1
rcssmin==1.1.2
redis==5.2.1
referencing==0.36.2
reportlab==4.3.1
requests==2.31.0
requests-cache==1.2.1
requests-toolbelt==1.0.0
rich==14.0.0
rjsmin==1.2.2
rpds-py==0.24.0
rsa==4.9
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.31.0
selenium-stealth==1.0.6
selenium-wire==5.1.0
six==1.17.0
sniffio==1.3.1
socks==0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
stem==1.8.2
svgpathtools==1.7.0
svgwrite==1.4.3
tenacity==9.1.2
threadpoolctl==3.6.0
torrequest==0.1.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
twilio==9.4.3
typing_extensions==4.12.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
undetected-chromedriver==3.5.5
uritemplate==4.1.1
url-normalize==2.2.1
urllib3==2.3.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wget==3.2
whitenoise==6.8.2
worldmap==0.1.7
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
